Added the file
This python code uses Tkinter and classes to build a GUI which gives user live weather data about any city in the World. 
The code was originally taken from this website: https://www.geeksforgeeks.org/python-real-time-weather-detection-using-tkinter/
However, extensive modifications have been done to the code to use classes for this event based GUI - which is ideal, since it gives access to all the variables in the best way for other uses such as making a plot.